+++ core/src/aggregator.rs
@@ -1,22 +1,21 @@
-use async_trait::async_trait;
-use bitcoin::{address::NetworkUnchecked, Address, OutPoint, Transaction};
-use secp256k1::schnorr;
-
 use crate::{
+    actor::Actor,
     config::BridgeConfig,
     errors::BridgeError,
     musig2::{
-        aggregate_nonces, AggregateFromPublicKeys, MuSigAggNonce, MuSigPartialSignature,
-        MuSigPubNonce,
+        aggregate_nonces, aggregate_partial_signatures, AggregateFromPublicKeys, MuSigAggNonce,
+        MuSigPartialSignature, MuSigPubNonce,
     },
     traits::rpc::AggregatorServer,
     transaction_builder::TransactionBuilder,
-    utils::{
-        aggregate_move_partial_sigs, aggregate_operator_takes_partial_sigs,
-        aggregate_slash_or_take_partial_sigs, handle_taproot_witness_new,
-    },
+    utils::{self, handle_taproot_witness_new},
     EVMAddress, UTXO,
 };
+use async_trait::async_trait;
+use bitcoin::hashes::Hash;
+use bitcoin::{address::NetworkUnchecked, Address, OutPoint, Transaction};
+use bitcoincore_rpc::RawTx;
+use secp256k1::schnorr;
 
 #[derive(Debug, Clone)]
 pub struct Aggregator {
@@ -36,6 +35,167 @@ impl Aggregator {
             config,
             nofn_xonly_pk,
         })
+    }
+
+    fn aggregate_slash_or_take_partial_sigs(
+        &self,
+        deposit_outpoint: OutPoint,
+        kickoff_utxo: UTXO,
+        operator_xonly_pk: secp256k1::XOnlyPublicKey,
+        operator_idx: usize,
+        agg_nonce: &MuSigAggNonce,
+        partial_sigs: Vec<[u8; 32]>,
+    ) -> Result<[u8; 64], BridgeError> {
+        let musig_agg_xonly_pubkey_wrapped = secp256k1::XOnlyPublicKey::from_musig2_pks(
+            self.config.verifiers_public_keys.clone(),
+            None,
+            false,
+        );
+        let mut tx = TransactionBuilder::create_slash_or_take_tx(
+            deposit_outpoint,
+            kickoff_utxo,
+            &operator_xonly_pk,
+            operator_idx,
+            &musig_agg_xonly_pubkey_wrapped,
+            self.config.network,
+        );
+        tracing::debug!("SLASH_OR_TAKE_TX: {:?}", tx);
+        tracing::debug!("SLASH_OR_TAKE_TX weight: {:?}", tx.tx.weight());
+        let message: [u8; 32] = Actor::convert_tx_to_sighash_script_spend(&mut tx, 0, 0)
+            .unwrap()
+            .to_byte_array();
+        tracing::debug!("aggregate SLASH_OR_TAKE_TX message: {:?}", message);
+        let final_sig: [u8; 64] = aggregate_partial_signatures(
+            self.config.verifiers_public_keys.clone(),
+            None,
+            false,
+            agg_nonce,
+            partial_sigs,
+            message,
+        )?;
+        tracing::debug!("aggregate SLASH_OR_TAKE_TX final_sig: {:?}", final_sig);
+        tracing::debug!(
+            "aggregate SLASH_OR_TAKE_TX for verifiers: {:?}",
+            self.config.verifiers_public_keys.clone()
+        );
+        tracing::debug!(
+            "aggregate SLASH_OR_TAKE_TX for operator: {:?}",
+            operator_xonly_pk
+        );
+        Ok(final_sig)
+    }
+
+    fn aggregate_operator_takes_partial_sigs(
+        &self,
+        deposit_outpoint: OutPoint,
+        kickoff_utxo: UTXO,
+        operator_xonly_pk: &secp256k1::XOnlyPublicKey,
+        operator_idx: usize,
+        agg_nonce: &MuSigAggNonce,
+        partial_sigs: Vec<[u8; 32]>,
+    ) -> Result<[u8; 64], BridgeError> {
+        let nofn_xonly_pk = secp256k1::XOnlyPublicKey::from_musig2_pks(
+            self.config.verifiers_public_keys.clone(),
+            None,
+            false,
+        );
+
+        let move_tx_handler = TransactionBuilder::create_move_tx(
+            deposit_outpoint,
+            &EVMAddress([0u8; 20]),
+            Address::p2tr(
+                &utils::SECP,
+                *utils::UNSPENDABLE_XONLY_PUBKEY,
+                None,
+                self.config.network,
+            )
+            .as_unchecked(),
+            &nofn_xonly_pk,
+            self.config.network,
+        );
+        let bridge_fund_outpoint = OutPoint {
+            txid: move_tx_handler.tx.compute_txid(),
+            vout: 0,
+        };
+        let slash_or_take_tx_handler = TransactionBuilder::create_slash_or_take_tx(
+            deposit_outpoint,
+            kickoff_utxo,
+            operator_xonly_pk,
+            operator_idx,
+            &nofn_xonly_pk,
+            self.config.network,
+        );
+        let slash_or_take_utxo = UTXO {
+            outpoint: OutPoint {
+                txid: slash_or_take_tx_handler.tx.compute_txid(),
+                vout: 0,
+            },
+            txout: slash_or_take_tx_handler.tx.output[0].clone(),
+        };
+        let mut tx_handler = TransactionBuilder::create_operator_takes_tx(
+            bridge_fund_outpoint,
+            slash_or_take_utxo,
+            operator_xonly_pk,
+            &nofn_xonly_pk,
+            self.config.network,
+        );
+        tracing::debug!(
+            "OPERATOR_TAKES_TX with operator_idx:{:?} {:?}",
+            operator_idx,
+            tx_handler.tx
+        );
+        tracing::debug!("OPERATOR_TAKES_TX_HEX: {:?}", tx_handler.tx.raw_hex());
+        tracing::debug!("OPERATOR_TAKES_TX weight: {:?}", tx_handler.tx.weight());
+        let message: [u8; 32] = Actor::convert_tx_to_sighash_pubkey_spend(&mut tx_handler, 0)
+            .unwrap()
+            .to_byte_array();
+        let final_sig: [u8; 64] = aggregate_partial_signatures(
+            self.config.verifiers_public_keys.clone(),
+            None,
+            true,
+            agg_nonce,
+            partial_sigs,
+            message,
+        )?;
+        tracing::debug!("OPERATOR_TAKES_TX final_sig: {:?}", final_sig);
+        Ok(final_sig)
+    }
+
+    fn aggregate_move_partial_sigs(
+        &self,
+        deposit_outpoint: OutPoint,
+        evm_address: &EVMAddress,
+        recovery_taproot_address: &Address<NetworkUnchecked>,
+        agg_nonce: &MuSigAggNonce,
+        partial_sigs: Vec<[u8; 32]>,
+    ) -> Result<[u8; 64], BridgeError> {
+        let musig_agg_xonly_pubkey_wrapped = secp256k1::XOnlyPublicKey::from_musig2_pks(
+            self.config.verifiers_public_keys.clone(),
+            None,
+            false,
+        );
+        let mut tx = TransactionBuilder::create_move_tx(
+            deposit_outpoint,
+            evm_address,
+            recovery_taproot_address,
+            &musig_agg_xonly_pubkey_wrapped,
+            self.config.network,
+        );
+        // println!("MOVE_TX: {:?}", tx);
+        // println!("MOVE_TXID: {:?}", tx.tx.compute_txid());
+        let message: [u8; 32] = Actor::convert_tx_to_sighash_script_spend(&mut tx, 0, 0)
+            .unwrap()
+            .to_byte_array();
+        let final_sig: [u8; 64] = aggregate_partial_signatures(
+            self.config.verifiers_public_keys.clone(),
+            None,
+            false,
+            agg_nonce,
+            partial_sigs,
+            message,
+        )?;
+
+        Ok(final_sig)
     }
 
     pub async fn aggregate_pub_nonces(
@@ -65,10 +225,9 @@ impl Aggregator {
     ) -> Result<Vec<schnorr::Signature>, BridgeError> {
         let mut slash_or_take_sigs = Vec::new();
         for i in 0..partial_sigs[0].len() {
-            let agg_sig = aggregate_slash_or_take_partial_sigs(
+            let agg_sig = self.aggregate_slash_or_take_partial_sigs(
                 deposit_outpoint,
                 kickoff_utxos[i].clone(),
-                self.config.verifiers_public_keys.clone(),
                 self.config.operators_xonly_pks[i],
                 i,
                 &agg_nonces[i + 1 + self.config.operators_xonly_pks.len()].clone(),
@@ -76,7 +235,6 @@ impl Aggregator {
                     .iter()
                     .map(|v| v.get(i).cloned().unwrap())
                     .collect::<Vec<_>>(),
-                self.config.network,
             )?;
 
             slash_or_take_sigs.push(secp256k1::schnorr::Signature::from_slice(&agg_sig)?);
@@ -93,15 +251,13 @@ impl Aggregator {
     ) -> Result<Vec<schnorr::Signature>, BridgeError> {
         let mut operator_take_sigs = Vec::new();
         for i in 0..partial_sigs.len() {
-            let agg_sig = aggregate_operator_takes_partial_sigs(
+            let agg_sig = self.aggregate_operator_takes_partial_sigs(
                 deposit_outpoint,
                 kickoff_utxos[i].clone(),
                 &self.config.operators_xonly_pks[i].clone(),
                 i,
-                self.config.verifiers_public_keys.clone(),
                 &agg_nonces[i + 1].clone(),
                 partial_sigs.iter().map(|v| v[i]).collect(),
-                self.config.network,
             )?;
 
             operator_take_sigs.push(secp256k1::schnorr::Signature::from_slice(&agg_sig)?);
@@ -117,14 +273,12 @@ impl Aggregator {
         agg_nonce: MuSigAggNonce,
         partial_sigs: Vec<MuSigPartialSignature>,
     ) -> Result<Transaction, BridgeError> {
-        let agg_move_tx_final_sig = aggregate_move_partial_sigs(
+        let agg_move_tx_final_sig = self.aggregate_move_partial_sigs(
             deposit_outpoint,
             &evm_address,
             &recovery_taproot_address,
-            self.config.verifiers_public_keys.clone(),
             &agg_nonce,
             partial_sigs,
-            self.config.network,
         )?;
 
         let move_tx_sig = secp256k1::schnorr::Signature::from_slice(&agg_move_tx_final_sig)?;
